+++ src/aosm/azext_aosm/custom.py
@@ -9,29 +9,29 @@ from azext_aosm.cli_handlers.onboarding_cnf_handler import OnboardingCNFCLIHandl
 from azext_aosm.cli_handlers.onboarding_vnf_handler import OnboardingVNFCLIHandler
 from azext_aosm.cli_handlers.onboarding_nsd_handler import OnboardingNSDCLIHandler
 from azext_aosm.common.command_context import CommandContext
-from azext_aosm.common.constants import ALL_PARAMETERS_FILE_NAME
+from azext_aosm.common.constants import ALL_PARAMETERS_FILE_NAME, CNF, VNF
 from azure.cli.core.commands import AzCliCommand
 from azure.cli.core.azclierror import UnrecognizedArgumentError
 
 
 def onboard_nfd_generate_config(definition_type: str, output_file: str | None):
     """Generate config file for onboarding NFs."""
-    if definition_type == "cnf":
+    if definition_type == CNF:
         handler = OnboardingCNFCLIHandler()
         handler.generate_config(output_file)
-    elif definition_type == "vnf":
+    elif definition_type == VNF:
         handler = OnboardingVNFCLIHandler()
         handler.generate_config(output_file)
     else:
         raise UnrecognizedArgumentError("Invalid definition type")
 
 
-def onboard_nfd_build(definition_type: str, config_file: Path):
+def onboard_nfd_build(definition_type: str, config_file: Path, skip: str = None):
     """Build the NF definition."""
-    if definition_type == "cnf":
-        handler = OnboardingCNFCLIHandler(Path(config_file))
+    if definition_type == CNF:
+        handler = OnboardingCNFCLIHandler(Path(config_file), skip=skip)
         handler.build()
-    elif definition_type == "vnf":
+    elif definition_type == VNF:
         handler = OnboardingVNFCLIHandler(Path(config_file))
         handler.build()
     else:
@@ -52,11 +52,15 @@ def onboard_nfd_publish(
             "definition_folder": Path(build_output_folder),
         },
     )
-    if definition_type == "cnf":
-        handler = OnboardingCNFCLIHandler(Path(build_output_folder,ALL_PARAMETERS_FILE_NAME))
+    if definition_type == CNF:
+        handler = OnboardingCNFCLIHandler(
+            Path(build_output_folder, ALL_PARAMETERS_FILE_NAME)
+        )
         handler.publish(command_context=command_context)
-    elif definition_type == "vnf":
-        handler = OnboardingVNFCLIHandler(Path(build_output_folder,ALL_PARAMETERS_FILE_NAME))
+    elif definition_type == VNF:
+        handler = OnboardingVNFCLIHandler(
+            Path(build_output_folder, ALL_PARAMETERS_FILE_NAME)
+        )
         handler.publish(command_context=command_context)
     else:
         raise UnrecognizedArgumentError("Invalid definition type")
@@ -88,7 +92,10 @@ def onboard_nsd_build(config_file: Path, cmd: AzCliCommand):
     handler.build()
 
 
-def onboard_nsd_publish(cmd: AzCliCommand, build_output_folder: Path, no_subscription_permissions: bool = False,
+def onboard_nsd_publish(
+    cmd: AzCliCommand,
+    build_output_folder: Path,
+    no_subscription_permissions: bool = False,
 ):
     """Publish the NF definition."""
     command_context = CommandContext(
@@ -98,7 +105,9 @@ def onboard_nsd_publish(cmd: AzCliCommand, build_output_folder: Path, no_subscri
             "definition_folder": Path(build_output_folder),
         },
     )
-    handler = OnboardingNSDCLIHandler(Path(build_output_folder, ALL_PARAMETERS_FILE_NAME))
+    handler = OnboardingNSDCLIHandler(
+        Path(build_output_folder, ALL_PARAMETERS_FILE_NAME)
+    )
     handler.publish(command_context=command_context)