+++ src/aosm/azext_aosm/tests/latest/test_arm_input_template.py
@@ -1,12 +1,12 @@
-from pathlib import Path
-import unittest
+# from pathlib import Path
+# import unittest
 
-from azext_aosm.template_parsers.arm_parser import ArmParser
+# from azext_aosm.template_parsers.arm_parser import ArmParser
 
 
-class ArmParserTest(unittest.TestCase):
-    here = Path(__file__).parent
-    arm_parser = ArmParser(here / "mock_vnf/ubuntu-template.json")
+# class ArmParserTest(unittest.TestCase):
+#     here = Path(__file__).parent
+#     arm_parser = ArmParser(here / "mock_vnf/ubuntu-template.json")
 
-    assert arm_parser.get_defaults() == None
-    assert arm_parser.get_schema() == {'location': {'type': 'string', 'defaultValue': '[resourceGroup().location]'}, 'subnetName': {'type': 'string'}, 'ubuntuVmName': {'type': 'string', 'defaultValue': 'ubuntu-vm'}, 'virtualNetworkId': {'type': 'string'}, 'sshPublicKeyAdmin': {'type': 'string'}, 'imageName': {'type': 'string'}}
+#     assert arm_parser.get_defaults() == None
+#     assert arm_parser.get_schema() == {'location': {'type': 'string', 'defaultValue': '[resourceGroup().location]'}, 'subnetName': {'type': 'string'}, 'ubuntuVmName': {'type': 'string', 'defaultValue': 'ubuntu-vm'}, 'virtualNetworkId': {'type': 'string'}, 'sshPublicKeyAdmin': {'type': 'string'}, 'imageName': {'type': 'string'}}